rename request state 'enabled' to 'started'
with the new 'stopped' state there was a possible confusion what each
state could do. Rename enabled to started to started to address this.

Enabled can still be used for backward compatibility and will be
removed on the next major release.

The TestHardware does not recognizes the "service enable" command
anymore, the regression test which made use of it were ported to
"service started" command.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>