is_node_shutdown: check for correct systemd targets
shutdown.target is active every time when the node shuts down, be it
reboot, poweroff, halt or kexec.
As we want to return true only when the node powers down without a
restart afterwards this was wrong.
Match only poweroff.target and halt.target, those two systemd targets
which cause a node shutdown without a reboot.

Enhance also the regular expression so that we do not falsely match
when a target includes poweroff.target in its name, e.g.
not-a-poweroff.target

Also pass the 'full' flag to systemctl to ensure that target name do
not get ellipsized or cut off.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>