Simulator: do not use cursor position to insert log
While the cursor is set invisible it can be still set by clicking
somewhere in the log window. As we uset insert_at_cursor() to append
text, asuming that the cursor is always at the end, this results in
an erroneous behavior where new log text get inserted at arbitrary
positions.
Use get_end_iter() to get the real end of the text buffer instead.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>