Simulator: fix scrolling to end of cluster log view
As the cursor may be set by the user, even if invisible, scrolling to
its position to reach the end of the TextView is potentially flawed.
Connect a signal on the 'size-allocate' event of the TextView element
instead. This event gets triggered on all size changes of the
TextViews.
In the call back use the ScrolledWindow methods to correctly scroll
to the end of its child (the TextView).

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>