sim: improve canceling the migrate dialog
We could only cancel the migrate dialog by pressing ESC or - if the
used window manager supports it - by pressing the windows "X" button
in the window border.
Pressing ESC caused a warning because the result of the domain was
now a string containing the signal names, as we checked for an
integer to see if the "Ok" button was pressed perl warned us about:
 > Argument "closed" isn't numeric in int at ...

Improve this by adding a cancel button and by switching the button
return values from integer to strings, which can be compared in a
more general way.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>