lrm.service: do not timeout on stop
we must shut all services down when stopping the LRM for a host
shutdown, this can take longer than 95 seconds and should not
get interrupted to ensure a gracefull poweroff.

The watchdog is still active untill all services got stopped so we
still are safe from a freeze or equivalent failure.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>