postinst: use auto generated postinst
This was introduced for cleaning up an possible left over systemd
watchdog mux enable link, which is gone for good now.

Then it was extended with trigger targets, as the HA Manager services
now restart when the pve-api-update trigger fires.
As the autogenerated postinst does the same unconditionally for the
pve-ha-lrm.service and pve-ha-crm.service already we may remove it
too.
The only difference is that try-restart is used by the auto generated
script, not reload-or-try-restart, but this does not matter, as the
HA services have currently no reload ability.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>