manager: select service node: allow to force best-score selection withot try-next
useful for re-balanacing on start, where we do not want to exclude
the current node like setting the $try_next param does, but also
don't want to favor it like not setting the $try_next param does.

We might want to transform both, `try_next` and `best_scored` into a
single `mode` parameter to reduce complexity and make it more
explicit what we want here.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>