manager: service start: make EWRONG_NODE a non-fatal error
traverse the usual error counting mechanisms, as then the
select_service_node helper either picks up the right node and it
starts there or it can trigger fencing of that.

Note, in practice this normally can only happen if the admin
butchered around in the node cluster state, but as we only select the
safe nodes from the configured groups, we should be safe in any case.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>