sim env: derive service usage from ID as fallback
so that we don't need to specify all usage stats explicitly for
bigger tests.

Note, we explicitly use two digits for memory as with just one a lot
of services are exactly the same, which gives us flaky tests due to
rounding, or some flakiness in the rust code - so this is a bit of a
stop gap for that too and should be reduced to a single digit once
we fixed it in the future.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>