manager: ensure node-request state transferred to new active CRM
We do not just take the full CRM status of the old master if a new
one gets active, we only take over the most relevant parts like node
state. But the relative new node_request object entry is also
important, as without that a maintenance state request may get lost
if a new CRM becomes the active master.

Simply copy it over on initial manager construction, if it exists.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>