test behavior of maintenance mode with another shutdown policy
Encode what happens if a node is in maintenance and gets shutdown
with a shutdown policy other than 'migrate' (= maintenance mode)
active.

Currently it's causing disabling the maintenance mode and also might
make a fence even possible (if not all service got moved already).
This will be addressed in the next commit.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>